little changes in unit tests and exceptions added

+++ src/comment/comment.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateCommentDto } from './dto/create-comment.dto';
 import { UpdateCommentDto } from './dto/update-comment.dto';
@@ -10,61 +10,74 @@ export class CommentService {
   async findAll() {
     const comments = await this.prisma.comment.findMany()
 
-    if (comments.length === 0) {
-      return { error: 'Még nincsenek kommentek' }
-    }
-
     return comments
   }
 
   async findOne(id: number) {
-    return this.prisma.comment.findUnique({ where: { id } })
+    const comment = await this.prisma.comment.findUnique({ where: { id } })
+
+    if (!comment) {
+      throw new NotFoundException()
+    }
+
+    return comment
   }
 
   async findAllByUser(userID: number) {
     const comments = await this.prisma.comment.findMany({ where: { userID } })
 
     if (comments.length === 0) {
-      return { message: 'Még nincsenek kommentek' }
-    } else {
-      return comments
+      throw new NotFoundException("User did not post any comments!")
     }
-  }
 
-  /*async findOneByUser(userID:number) {
-        const comments = await this.prisma.comment.findUnique({ where: {userID} })
-    }*/
+    return comments
+  }
 
   async findAllByPlace(placeID: number) {
     const comments = await this.prisma.comment.findMany({
       where: { placeID: placeID },
-    });
+    })
 
     if (comments.length === 0) {
-      return { message: 'A helyhez még nincsenek kommentek' }
-    } else {
-      return comments
+      throw new NotFoundException("Place does not have any comments!")
     }
+
+    return comments
   }
 
   async add(data: CreateCommentDto) {
     return this.prisma.comment.create({ data })
   }
 
-  async remove(id: number) {
+  async remove(id: number, userId: number) {
+    const comment = await this.prisma.comment.findUnique({ where: { id } })
+
+    if (!comment) {
+      throw new NotFoundException("Can not delete comment, not found!")
+    }
+
+    if(comment.userID !== userId) {
+      throw new ForbiddenException("You can only delete your own comment!")
+    }
+
     return this.prisma.comment.delete({ where: { id } })
   }
 
   async update(id: number, data: UpdateCommentDto) {
+    const comment = await this.prisma.comment.findUnique({ where: { id } })
+
+    if (!comment) {
+      throw new NotFoundException("No comment found!")
+    }
+
     return this.prisma.comment.update({ where: { id }, data })
   }
+
   async findAllByGooglePlace(googlePlaceID: string) {
-    const place = await this.prisma.place.findFirst({
-      where: { googleplaceID: googlePlaceID },
-    })
+    const place = await this.prisma.place.findFirst({ where: { googleplaceID: googlePlaceID } })
 
     if (!place) {
-      return []
+      throw new NotFoundException("Place not found!")
     }
 
     return this.prisma.comment.findMany({
